cron build: Use local mirror, toss clone after
Instead of attempting to work from a standard repo clone,
update a local mirror and then generate a clone from that
mirror. This allows for potentially destructive operations
without any concern of bandwidth waste from a completely
fresh clone from the remote for each run of the cron
build script.

Now that we are creating a fresh clone for each run of the
cron script, we can update the script to remove steps
specific to working from a detached head state.

- Simplify "get branch" functions by specifically asking git
  for remote branches instead of all branches and then filtering
  for remote branches.

- Create new functions for returning the latest tag and
  the latest stable version.